+++ .github/CHANGELOG.md
@@ -1,8 +1,14 @@
 ## Changelog
 
+Update <small>_ March 2025</small>
+
+- fix: Update FAQ links (13/03/2025)
+- feat: rule update (13/03/2025)
+
 Update <small>_ November 2024</small>
 
-- feat: generic wrapper for JSON APIs (01/11/2024)
+- feat: Update welcome message for A380X release (09/11/2024)
+- fix: TS v5.5 breaks build (09/11/2024)
 
 Update <small>_ October 2024</small>